Edit user and router done

+++ tools/drop_db.js
@@ -1,7 +1,7 @@
 var dbs = db.getMongo().getDBNames()
 for(var i in dbs){
     db = db.getMongo().getDB( dbs[i] );
-    if (db.getName() != 'admin') {
+    if (db.getName() != 'admin' && db.getName() != 'prosper-cr') {
     	print( "dropping db " + db.getName() );
     	db.dropDatabase();
 	}
